third_party/WebKit/LayoutTests/fast/html/imports/import-readd.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel=import href="./resources/style-red.html">
</head>
<body>
<h1 class="target">Should be red</h1>
<script>
var t = async_test('Styles are re-applied after removed imports are re-added');
window.onload = function() {
    t.step(function() {
        var importLink = document.querySelector('link');
        document.head.removeChild(importLink);
        document.head.appendChild(importLink);
        assert_equals(window.getComputedStyle(document.querySelector('.target')).color, 'rgb(255, 0, 0)');
        Array.prototype.forEach.call(document.querySelectorAll("h1"), function(e) { e.style.display = "none"; });
        t.done();
    });
}
</script>
</body>
</html>
